[flatland] Remove locks around VkRenderer::ExtractImage()
VkRenderer::ExtractImage() is a potentially expensive method
called from a worker thread. This CL adds threading annotations
to ensure it is never locked, and removes the current locks around
it.

Also found a spot where we make an unnecessary call to sysmem (we
were checking that a buffer was allocated in ImportReadbackImage()
even though we'd already checked the same thing before the call) and
removed it.

Fixed: 119195
Change-Id: I6461d1946c99599670409519d03843eadb5e13c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790518
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>