[forensics] Temporarily disable hwasan for exception_broker_integration_test
This test propagates a bunch of tagged pointers to crashpad which
propagates them to fuchsia syscalls which do not accept tags. We should
come back and addresses these once we know if we want crashpad to accept
tagged pointers. For now we can just disable it to unblock migrating the
hwasan FYI bot to CI.

Bug: 118712
Change-Id: Ifb5aa74bc45f099f4dcd09d4ec69121ecbb45ec4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790725
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>