[musl] Clean up usage of ZX_HANDLE_FIXED_BITS_MASK
Code in MUSL was attempting to make a unique identifier out of a
handle, but to make sure that the MSB of the handle was not set in the
process.  As a result, it wanted to shift the handle value such that
the lowest fixed bits of the handle were shifted away to the right,
guaranteeing that the MSB of the generated identifier was guaranteed
to be 0.

ZX_HANDLE_FIXED_BITS_MASK specifies which bits of handles are
guaranteed to be fixed, and currently has the value `3`.  Code in MUSL
was computing the amount to shift by as `ZX_HANDLE_FIXED_BITS_MASK - 1`.

This just _happens_ to work when the low 2 bits are fixed, but it
clearly does not always work.  For example, if the 3 low bits were
fixed, the mask would be 0x7, and MASK - 1 == 6 (not 3).
Additionally, the fixed bit mask could (someday) even change such that
the fixed bits in the word were not even contiguous.

What is _actually_ important here is that the low bit of the handle
value is fixed and can be discarded without ruining the
process-wide-unique property of the shifted handle.  So, instead of
shifting by an amount which is a function of the fixed bits mask,
instead just shift by 2, and `static_assert` that the two low bits are
members of the fixed mask.  As long as this never changes (it never
should), we will be OK.

Change-Id: I9bfdc596175df4ecceb60d4b440d9214114d29b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715975
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>