[iwlwifi] Create a new dispatcher for FIDL server in wlanphy-impl-device
This CL partially reverts fxr/761123.

wlan-softmac-device and wlanphy-impl-device are in the same driver host
and share the same default driver dispatcher, the device unbind and
remove tasks are all scheduled on this dispatcher. Thus, the
sync_completion_wait in CreateIface() in wlanphy-impl-device will block
the tasks on this dispatcher including unbind and remove tasks. So no
matter how long we wait in CreateIface(), the device unbind/remove
tasks will wait for the completion of CreateIface(), and CreateIface()
always completes as a timeout since it waits for the tasks that wait
for it.

This change creates a separate dispatcher to do the job.

Bug: 118079
Test: Delete iface 0 with "wlan-dev iface del 0" and no interference
observed. New iface is created successfully.

Change-Id: I1c06cebb2ea40a0b184858b35930613fe18b8bdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781589
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Louis Lo <yjlou@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>