[iwlwifi] Reset wait_for_delete completion after signaling it
"wait_for_delete" completion is used by phy_create_iface() for waiting
for in-progress iface deletion. When an iface finishes the deletion,
"wait_for_delete" will be signaled, but since it's not reset, it won't
be ready to be waited on for the next phy_create_iface(). This change
resets this completion after signaling it.

Bug: 118079
Test: Manually deleted iface for 5 times with "wlan-dev iface del [id]",
and all of them got recovered successfully.

Change-Id: I6a0962f05d6805d9a55e97b571bf4968a3a2971a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783895
Reviewed-by: Louis Lo <yjlou@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>