[chromium] Replace test deps on web_engine with new uses_web_engine target
This provides a clearer indication of the reason for the dependency and
will allow configuring this target differently if necessary.
Also add TODOs related to general changes to such deps.

Add WebEngine to Workstation's buildbot bundle so Workstation bots
can run the integration tests that require it.

Bug: 119075, 119212
Change-Id: Iaefdaabde9914bf1ae406e58116a0acb4ba5a8f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788026
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>