[elf_runner] Use smaller buffers to forward stdio to logs.
Limits the size of the forwarding buffer to the number of bytes
which the kernel says a socket has to read, up to the maximum
message size that we'll forward to logs. While at times this may
grow to as much as 50% of the size of the previous implementations'
buffers, we also use fresh heap allocations for the internal buffer
each time a message is yielded so that large allocations will be
quickly freed after being forwarded to the logs.

Adds a new unit test which replicates the behavior expected by
some intl examples' integration tests.

Results in a 500-800KiB reduction in heap usage by Component
Manager in the product configurations I tested. While the previous
implementation allocated ~10MiB of buffers in some configurations,
most of the pages for those allocations were never touched and thus
never actually provisioned by the vmm. That said, the previous
implementation still had the potential to allocate all of those pages
in pathological scenarios, so in addition to the savings these
smaller buffers should also result in more predictable memory
usage.

Fixed: 118020
Run-All-Tests: true

Change-Id: I03d23229ea1c05fb836f10f3518eefde10c1c4a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782162
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>