[superproject] Roll cobalt Remove CHECK failure if FileObservationStore can't create a directory.
This is consistent with the other uses of MakeDirectory in the
ImmediateLocalAggregateStorage. All the other uses of the directory
and file system in the FileObservationStore seem to fail gracefully.
So this should only result in dropping observations, and not some
other failure downstream of the CHECK.

Original-Bug: b/265935870
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/791164
Original-Revision: 44b782c5734825fab4ad267a59c49860cc9e6362
GitOrigin-RevId: 99b92eb901c881075339fdd275f5d63ad35dd26c
Roller-URL: https://ci.chromium.org/b/8791494099934573345
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: If8632a5d2363fe022ff36e68723eb7f4fa60aa3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791810