[build] Remove todo regarding investigating asan odr indicator
This has been investigated. The asan odr checkecing requires
default visibility symbols for the odr indicators to check
if the indicators have been interposed upon. This implies that
a symbol of the same name was found at runtime.

It's that clear that we will ever want to turn
-fsanitize-address-use-odr-indicator on. Having multiple definitions
of a symbol across shared objects is well defined in ELF.

Bug: 114037

Change-Id: I4b8e557dab97944acb3cb7101fdee9d3b8d6e487
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791348
Reviewed-by: Leonard Chan <leonardchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>