[fxfs] Enable LTO
This gains back the performance which was lost when splitting Fxfs into
several crates.  See, for example, WalkDirectoryTreeWarm.

The cost is a ~45% slowdown in compilation time for fxfs (19s -> 33s on
my machine).  If this ends up being problematic for productivity, we
could further split up Fxfs into a binary variant used for tests (which
won't have LTO) and another used for benchmarks and production (which
would have LTO), but that doesn't seem necessary given the build time is
still reasonable.

Note that LTO is not applied to the unit-test suite, so building the
unit-test suite won't be made slower.

Bug: 119533

Change-Id: I82d7fb855c76332f34f48047f5c7f60d964029b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791112
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>