[versioning] Expose the target API level to non Fuchsia C++ targets
This exposes the target API level value to the non Fuchsia C++
targets (e.g. the host tools). This remove the definition that existed
in the debug codebase to make it usable more globally (and it also fix
the fact that this define was tracking the wrong value).

Change-Id: Iabfeaf76d5dd19352aaf1806f4fd39744843cbc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791523
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>