[ffx] Don't include any ffx command in subtool re-run
For some reason this was trying to include part of the ffx command line
in the re-run command for a discovered subtool. This meant that if you
ran it with `fx ffx`, it would trim off just the `fx`, but not the `ffx`
and produce a nonsensical command line like `ffx-repository ffx
repository`.

I'm honestly not sure why this was here, and it doesn't make sense.
Removing it makes that work again.

Fixed: 120315
Change-Id: Icefff185bc1949e6aebc286872c7811f2aec0ec0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791957
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>