[sdk] Remove fuchsia.modular.session from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Bug: 120471
Change-Id: I79d12d326e4ceabb2d878efae34e46ce9c8476d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792365
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>