[ffx][errors] Don't print "more info" on user err
This was likely a regression from the old top level error handler code.
We don't really want to tell the user there's more info somewhere else,
since the tool developer probably put relevant information into the user
error to begin with.

This also corrects an issue where we were printing the wrong level of
error, which caused us to drop the detailed error info we were trying to
output along with the `BUG: ` line unexpected errors are supposed to be
prefixed with.

Fixed: 120269
Change-Id: I1bd3eb1370de59fff56e8c5816d8ff31118ff361
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791635
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Bob Evans <bobevans@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>