[netstack3] Support non-Vec insertion target
For TCP, there can be at most one listening socket on an address, which
means that an address entry should hold its listening socket in an
Option. In order to support SO_REUSEADDR, the TCP listener state for an
address will need to support inserting a listening socket when there are
already bound sockets present. That means the Option for the listener
will need to be the target of an insertion, which means the &mut Vec
target is no longer general enough.

This CL cleans up what was admittedly a cut corner by putting the &mut
Vec behind a trait that allows only a single insertion. This also opens
up the door for non-Vec implementations, which will be needed for future
CLs.

Bug: 120272
Change-Id: Id90ffa09e40eda9fe04747105920b5011d595b33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792380
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>