[sdk] Remove fuchsia.cobalt from the partner SDK
This library is currently in the partner SDK category but it's not
exported in the SDK because it's not listed in
`//sdk/fidl:sdk_fidl_targets`. We could just add it to the list to
fix this but the fact that it's not exported means that it's
currently not used by the partners.

Bug: 120471
Change-Id: I206a33486d2ee28d8d98451f6742f88bebdd52bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792622
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>