[libc] Remove some __* symbols from the abi.
It seems likely that these symbols should have been hidden and
were used at one point as a way for libc to use the internal
symbol without fear of using the potentially interposed public
one. These symbols are not referenced from within libc, or in the
rest of Fuchsia's tree. Though we will do an lsc + wait the requisite
time before pushing.

I'm not sure if strdup should be weak or not. It is in glibc, but
it seems they are very liberal with making things weak. Note, the
distinction here will matter if fxrev.dev/772922 lands.

This will be useful when transitioning to llvm-libc string functions
because their implementation does not export these symbols.

Bug: 62780
Change-Id: I46dce9939541c130cb788dd2643f465bcddbf07c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788102
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Alex Brachet <abrachet@google.com>