[bt][fastpair] Report component metrics to Cobalt
Offer the Cobalt service from Cobalt to the Bluetooth-core realm.
Because bt-core is currently static, this definition exists in core.cml.
This will be removed when fxbug.dev/94055 is addressed and the
bluetooth-core realm is product-variable.

Route the Cobalt service to the Fast Pair component. This service is
marked as optional because it is not critical to component success. The
Fast Pair component is resilient to the unavailability of the
capability.

Log metrics occurrences any time a Fast Pair pairing completes
successfully or unsuccessfully.

Log metrics occurrences any time a Fast Pair request is received from
the Seeker.

Log the total Fast Pair pairing time to Cobalt.

File fxbug.dev/119634 to track logging Device Action requests to Cobalt.
Currently, Device Actions are considered Errors during parsing. Instead,
we should extract the relevant DA fields, log it, and record to metrics.

Bug: 114083
Test: fx test bt-fastpair-provider-tests
Change-Id: I2668c7dd16857e482a08660bbf723b9b13705a90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/769051
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>