[sdk] Enforce that all partner APIs are in `sdk_fidl_targets`
There's currently no mechanism to enforce that all FIDL libraries
labelled as "partner" get listed in //sdk/fidl:sdk_fidl_targets. The
consequence of this is that some APIs might be labelled as partner
without actually being in the partner SDK (fxb/120471) , while some
other only end up in the SDK because there's another library in that
list that depends on them.

Some cleanup has recently been done to ensure that all partner APIs are
tracked in this list, this CL adds an allowlist mechanism to prevent
this from regressing. One side effect of this is that it'll make it
simpler for developers to add their API to the partner SDK as they won't
have to know which list to update, the build system will tell them.

Bug: 120472
Change-Id: I634663c734e4351da54bf47e5e807a5a1a9727e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792644
Commit-Queue: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Shai Barack <shayba@google.com>