[dfv2] Refactor adding Node offers
Pull out the logic of turning an offer from the driver into an offer
that will be sent to the component framework.

This will make it easier to update this logic to not rely on mutable
tables.

Bug: 120602
Change-Id: I5ea5253905c7f6412266b702aaf27e659d6863bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792983
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>