[kernel][vm] Make sign conversion explicit.
This CL adds a static cast to make an explicit sign & truncation
conversion. When we do `~kLoanedStateIsLoaned` it gets converted to an
int and its actual value becomes `0xfffffffd` (signed), and then it gets
downcasted to `uint8_t` again.

This CL fixes ubsan checks. These are not strictly undefined behavior,
but ubsan reports them regardless because most of the time the effect is
unintended.

Bug: 32059
Change-Id: Ie6e07f04b183b493e70f92940fb95c33c99dc175
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792842
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>