[libdriver-integration-test] Remove thread and multiplexing
Make a separate async connection to the controller instead of
spawning a new thread.

This has a good chance of fixing a flake where the device was
torn down, because now the controller call is happening on the
same thread, meaning it is synchronized to the other calls in the test.

Bug: 119548
Multiply: libdriver-integration-test
Change-Id: Ie89a909a35b57f6ed5dcf3a69f1da1aae08a1648
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793082
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>