[kernel][dlog] Fix alignment of DLog::data_
mvanotti@ pointed out that UBSan noticed some unaligned access of
objects in data_.  This field is an array of bytes that holds records
that start with a dlog_header_t object so we must ensure it's aligned
as 4 or more to avoid undefined behavior.

Update two debuglog test cases to memcpy out of the fifo to avoid the
possibility of unaligned access.

Bug: 105063
Change-Id: I8fbda2e03e2da5ffb9f21d9be587665bf578be9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791804
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>