[fidl][cpp20] Support ObjectView<VectorView<MoveOnly>>
Consider the following code snippet:

    fidl::Arena;
    auto vv = fidl::VectorView<Handle>::FromExternal(...);
    fidl::ObjectView ov(arena, vv);

One would expect it to allocate the vector view header itself on the
arena, and then point the body to the content that was borrowed via
FromExternal.

However, it fails to compile in C++20. This is because `std::span` in
C++20 is implicitly constructed from any RangeType as long as said type
contains methods such as begin/end/etc (the precise detail is encoded in
a Concept). ObjectView then unnecessarily ask VectorView to copy its
contents into the arena, which also fails because Handle is move-only.

There are a number of ways to avoid that; this CL chooses the one that
seems least special-casy which is to re-order the priorities of
constructor calls in
sdk/lib/fidl/cpp/wire/include/lib/fidl/cpp/wire/arena.h.

ObjectView will first attempt to construct without providing the arena,
falling back to providing the arena if the former is not possible.

By itself, that change implies that the following code:

    char buffer[kSize];
    strlcpy(buffer, "hello", kSize);
    fidl::Arena arena;
    auto table = test::wire::SampleTable::Builder(arena).s(buffer).Build();

will now unsafely reference |buffer|. That's not good. The implicit
|StringView| from a string literal should only be used for string
constants (static storage duration). But people have been using it to
borrow their local/member variables. To ensure that |buffer| is still
copied, I added a SFINAE to require that the string passed to
StringView is a const char[], as opposed to any char buffer. It doesn't
fully avoid all unsafe references, but as can be seen in the user code
changes, it catches a lot of unintended uses, and moves them to
FromExternal.

It also means the following snippet

    fidl::Arena;
    auto builder = MyWireTable::Builder(arena)
        .some_string("hello")
	.Build();

will now allocate the table header and string
view header on the arena, then point the string view to the static
string literal "hello", instead of copying "hello" unnecessarily into
the arena. Overall this should make for more intuitive behavior.

RELNOTES: fidl::StringView(some_string) only works for string constants,
not arbitrary char buffers. If you get a compiler error, change it to
fidl::StringView::FromExternal(some_string) to indicate that you are
unsafely borrowing some_string.

TEST: Successfully built `src/lib/fidl/llcpp/tests` under C++20.

Bug: 113689
Change-Id: I470d1452d0ee9a1bec80f93546468dda28b24e63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792516
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Ian McKellar <ianloic@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>